+++ abrechnung/http/routers/common.py
@@ -1,7 +1,9 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, Depends
 from pydantic import BaseModel
 
 from abrechnung import MAJOR_VERSION, MINOR_VERSION, PATCH_VERSION, __version__
+from abrechnung.config import Config, ServiceConfig, ServiceMessage
+from abrechnung.http.dependencies import get_config
 
 router = APIRouter(
     prefix="/api",
@@ -34,3 +36,21 @@ async def get_version():
         "minor_version": MINOR_VERSION,
         "patch_version": PATCH_VERSION,
     }
+
+
+class FrontendConfig(BaseModel):
+    messages: list[ServiceMessage] | None = None
+    imprint_url: str | None = None
+    source_code_url: str
+    issue_tracker_url: str
+
+
+@router.get("/config", response_model=FrontendConfig, operation_id="get_config")
+async def get_frontend_config(config: Config = Depends(get_config)):
+    cfg = FrontendConfig(
+        messages=config.service.messages,
+        imprint_url=config.service.imprint_url,
+        source_code_url=config.service.source_code_url,
+        issue_tracker_url=config.service.issue_tracker_url,
+    )
+    return cfg