+++ frontend/apps/web/src/components/accounts/BalanceTable.tsx
@@ -30,12 +30,12 @@ export const BalanceTable: React.FC<Props> = ({ group }) => {
         {
             field: "totalConsumed",
             headerName: t("balanceTable.totalConsumed"),
-            renderCell: renderCurrency(group.currency_symbol, "red"),
+            renderCell: renderCurrency(group.currency_symbol, -1),
         },
         {
             field: "totalPaid",
             headerName: t("balanceTable.totalPaid"),
-            renderCell: renderCurrency(group.currency_symbol, "green"),
+            renderCell: renderCurrency(group.currency_symbol, 1),
         },
         {
             field: "balance",