+++ frontend/apps/web/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from "./useFormatCurrency";
 export * from "./useIsSmallScreen";
 export * from "./useFormatDatetime";
+export * from "./useGetAmountColor";