+++ CHANGELOG.md
@@ -7,8 +7,9 @@
 **BREAKING CHANGES**
 
 - changed config structure to only include the reverse-proxy base_url once in the `api` section
+- removed frontend config via separate `config.json` in favour of configuration in the backend config yaml
 - drop python 3.10 support
-- remove react native mobile app in favor of mobile optimized PWA
+- remove react native mobile app in favour of mobile optimized PWA
 
 **Features**